Agent: implement #437 — Refactor: Extract MainWindow.axaml panels into UserControls (Phase 2)
