+++ CAP.Avalonia/Views/MainWindow.axaml.cs
@@ -1,11 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Input;
-using Avalonia.Interactivity;
 using CAP.Avalonia.Services;
 using CAP.Avalonia.ViewModels;
-using CAP.Avalonia.ViewModels.Library;
-using System.ComponentModel;
-using System.Linq;
 
 namespace CAP.Avalonia.Views;
 
@@ -26,21 +22,12 @@ public partial class MainWindow : Window
                 vm.RightPanel.RoutingDiagnostics.FileDialogService = vm.FileDialogService;
                 vm.ViewportControl.GetViewportSize = GetActualViewportSize;
 
-                // Wire up rename dialog for group templates
-                vm.LeftPanel.ComponentLibrary.ShowRenameDialogAsync = async (currentName) =>
-                {
-                    var dialog = new RenameDialog(currentName);
-                    return await dialog.ShowDialog<string?>(this);
-                };
-
                 // Wire up clipboard for RoutingDiagnostics
                 vm.RightPanel.RoutingDiagnostics.CopyToClipboard = async (text) =>
                 {
                     var clipboard = Clipboard;
                     if (clipboard != null)
-                    {
                         await clipboard.SetTextAsync(text);
-                    }
                 };
 
                 // Wire up clipboard for DimensionValidator
@@ -48,45 +35,11 @@ public partial class MainWindow : Window
                 {
                     var clipboard = Clipboard;
                     if (clipboard != null)
-                    {
                         await clipboard.SetTextAsync(text);
-                    }
-                };
-
-                // Wire up clipboard for ErrorConsole
-                vm.BottomPanel.ErrorConsole.CopyToClipboard = async (text) =>
-                {
-                    var clipboard = Clipboard;
-                    if (clipboard != null)
-                    {
-                        await clipboard.SetTextAsync(text);
-                    }
-                };
-
-                // Wire up auto-scroll: scroll to the newest entry when entries are added
-                vm.BottomPanel.ErrorConsole.PropertyChanged += (s, e) =>
-                {
-                    if (e.PropertyName == nameof(vm.BottomPanel.ErrorConsole.EntryCount) && ErrorConsoleListBox != null)
-                    {
-                        var items = ErrorConsoleListBox.ItemsSource;
-                        if (items is System.Collections.IList list && list.Count > 0)
-                        {
-                            ErrorConsoleListBox.ScrollIntoView(list[list.Count - 1]);
-                        }
-                    }
                 };
 
                 // Wire up GridSplitter resize events
                 SetupPanelResizing(vm);
-
-                // Wire up LeftPanel.SelectedGroupTemplate changes to update ListBox selections
-                vm.LeftPanel.PropertyChanged += (s, e) =>
-                {
-                    if (e.PropertyName == nameof(vm.LeftPanel.SelectedGroupTemplate))
-                    {
-                        UpdateGroupTemplateListBoxSelections(vm.LeftPanel.SelectedGroupTemplate);
-                    }
-                };
             }
         };
     }
@@ -292,15 +245,6 @@ public partial class MainWindow : Window
         DesignCanvasControl.InvalidateVisual();
     }
 
-    private void ZoomToFitButton_Click(object? sender, RoutedEventArgs e)
-    {
-        if (DataContext is MainViewModel vm)
-        {
-            var (width, height) = GetActualViewportSize();
-            vm.ZoomToFit(width, height);
-        }
-    }
-
     /// <summary>
     /// Gets the actual viewport size (visible area) independent of zoom level.
     /// Uses the DesignCanvas control's own layout bounds, which correctly excludes
@@ -327,143 +271,6 @@ public partial class MainWindow : Window
             return (windowWidth, windowHeight);
 
         return (1400, 900); // Last-resort default matching the initial window size
-    }
-
-    /// <summary>
-    /// Handles pointer entering a group template item (shows delete button).
-    /// </summary>
-    private void OnGroupItemPointerEntered(object? sender, PointerEventArgs e)
-    {
-        if (sender is Border border && border.DataContext is GroupTemplateItemViewModel itemVm)
-        {
-            itemVm.IsHovered = true;
-        }
-    }
-
-    /// <summary>
-    /// Handles pointer leaving a group template item (hides delete button).
-    /// </summary>
-    private void OnGroupItemPointerExited(object? sender, PointerEventArgs e)
-    {
-        if (sender is Border border && border.DataContext is GroupTemplateItemViewModel itemVm)
-        {
-            itemVm.IsHovered = false;
-        }
-    }
-
-    /// <summary>
-    /// Handles selection change in UserGroups ListBox.
-    /// Extracts the GroupTemplate from GroupTemplateItemViewModel and sets it in LeftPanel.
-    /// </summary>
-    private void OnUserGroupsSelectionChanged(object? sender, SelectionChangedEventArgs e)
-    {
-        if (DataContext is not MainViewModel vm) return;
-        if (sender is not ListBox listBox) return;
-
-        if (listBox.SelectedItem is GroupTemplateItemViewModel itemVm)
-        {
-            vm.LeftPanel.SelectedGroupTemplate = itemVm.Template;
-            // Clear PDK groups selection
-            ClearPdkGroupsSelection();
-        }
-        else if (listBox.SelectedItem == null)
-        {
-            // Only clear if this was triggered by user action, not by code
-            if (e.RemovedItems.Count > 0 && e.AddedItems.Count == 0)
-            {
-                vm.LeftPanel.SelectedGroupTemplate = null;
-            }
-        }
-    }
-
-    /// <summary>
-    /// Handles selection change in PdkGroups ListBox.
-    /// Extracts the GroupTemplate from GroupTemplateItemViewModel and sets it in LeftPanel.
-    /// </summary>
-    private void OnPdkGroupsSelectionChanged(object? sender, SelectionChangedEventArgs e)
-    {
-        if (DataContext is not MainViewModel vm) return;
-        if (sender is not ListBox listBox) return;
-
-        if (listBox.SelectedItem is GroupTemplateItemViewModel itemVm)
-        {
-            vm.LeftPanel.SelectedGroupTemplate = itemVm.Template;
-            // Clear user groups selection
-            ClearUserGroupsSelection();
-        }
-        else if (listBox.SelectedItem == null)
-        {
-            // Only clear if this was triggered by user action, not by code
-            if (e.RemovedItems.Count > 0 && e.AddedItems.Count == 0)
-            {
-                vm.LeftPanel.SelectedGroupTemplate = null;
-            }
-        }
-    }
-
-    private void ClearUserGroupsSelection()
-    {
-        if (UserGroupsListBox != null)
-        {
-            UserGroupsListBox.SelectedItem = null;
-        }
-    }
-
-    private void ClearPdkGroupsSelection()
-    {
-        if (PdkGroupsListBox != null)
-        {
-            PdkGroupsListBox.SelectedItem = null;
-        }
-    }
-
-    /// <summary>
-    /// Clears both user and PDK group selections. Called from MainViewModel.
-    /// </summary>
-    public void ClearAllGroupSelections()
-    {
-        ClearUserGroupsSelection();
-        ClearPdkGroupsSelection();
-    }
-
-    /// <summary>
-    /// Updates ListBox selections to match the given GroupTemplate.
-    /// Finds the corresponding GroupTemplateItemViewModel and selects it.
-    /// </summary>
-    private void UpdateGroupTemplateListBoxSelections(CAP_Core.Components.Creation.GroupTemplate? template)
-    {
-        if (DataContext is not MainViewModel vm) return;
-
-        if (template == null)
-        {
-            // Clear all selections
-            ClearAllGroupSelections();
-        }
-        else
-        {
-            // Find and select the matching item in UserGroups
-            var userItem = vm.LeftPanel.ComponentLibrary.UserGroups.FirstOrDefault(vm => vm.Template == template);
-            if (userItem != null)
-            {
-                if (UserGroupsListBox != null)
-                {
-                    UserGroupsListBox.SelectedItem = userItem;
-                }
-                ClearPdkGroupsSelection();
-                return;
-            }
-
-            // Find and select the matching item in PdkGroups
-            var pdkItem = vm.LeftPanel.ComponentLibrary.PdkGroups.FirstOrDefault(vm => vm.Template == template);
-            if (pdkItem != null)
-            {
-                if (PdkGroupsListBox != null)
-                {
-                    PdkGroupsListBox.SelectedItem = pdkItem;
-                }
-                ClearUserGroupsSelection();
-            }
-        }
     }
 
 }
