+++ CAP.Avalonia/Views/Panels/ToolbarPanel.axaml
@@ -1,0 +1,96 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:vm="using:CAP.Avalonia.ViewModels"
+             x:Class="CAP.Avalonia.Views.Panels.ToolbarPanel"
+             x:DataType="vm:MainViewModel">
+
+    <ScrollViewer HorizontalScrollBarVisibility="Auto"
+                  VerticalScrollBarVisibility="Disabled"
+                  Background="#2d2d2d"
+                  Height="40">
+        <StackPanel Orientation="Horizontal">
+            <!-- Mode Buttons -->
+            <TextBlock Text="Mode:" VerticalAlignment="Center" Margin="10,0,5,0" Foreground="White"/>
+            <Button Content="[S] Select" Command="{Binding SetSelectModeCommand}"
+                    Width="80" Margin="2" Background="#3d3d3d" Foreground="LightBlue"/>
+            <Button Content="[C] Connect" Command="{Binding SetConnectModeCommand}"
+                    Width="85" Margin="2" Background="#3d3d3d" Foreground="Orange"/>
+            <Button Content="[D] Delete" Command="{Binding SetDeleteModeCommand}"
+                    Width="80" Margin="2" Background="#3d3d3d" Foreground="Red"/>
+
+            <Separator Width="1" Background="Gray" Margin="10,5"/>
+
+            <!-- Zoom Controls -->
+            <Button Content="+" Command="{Binding ZoomInCommand}" Width="30" Margin="2"/>
+            <Button Content="-" Command="{Binding ZoomOutCommand}" Width="30" Margin="2"/>
+            <Button Content="[F] Fit" x:Name="ZoomToFitButton" Click="ZoomToFitButton_Click"
+                    Width="50" Margin="2" Background="#3d3d3d" ToolTip.Tip="Zoom to fit all components (F)"/>
+            <Button Content="1:1" Command="{Binding ResetZoomCommand}"
+                    Width="35" Margin="2" ToolTip.Tip="Reset zoom to 100%"/>
+
+            <Separator Width="1" Background="Gray" Margin="10,5"/>
+
+            <!-- Undo/Redo -->
+            <Button Content="↶" Command="{Binding UndoCommand}"
+                    Width="30" Margin="2" Background="#3d3d3d" ToolTip.Tip="Undo (Ctrl+Z)"/>
+            <Button Content="↷" Command="{Binding RedoCommand}"
+                    Width="30" Margin="2" Background="#3d3d3d" ToolTip.Tip="Redo (Ctrl+Y)"/>
+
+            <Separator Width="1" Background="Gray" Margin="10,5"/>
+
+            <!-- Actions -->
+            <Button Content="❌" Command="{Binding DeleteSelectedCommand}"
+                    Width="36" Margin="2" Background="#4d3d3d" ToolTip.Tip="Delete Selected (Del)"/>
+
+            <Separator Width="1" Background="Gray" Margin="10,5"/>
+
+            <!-- Group/Ungroup -->
+            <Button Command="{Binding CreateGroupCommand}"
+                    Width="85" Margin="2" Background="#3d3d3d" ToolTip.Tip="Create Group (Ctrl+G)">
+                <StackPanel Orientation="Horizontal" Spacing="4">
+                    <PathIcon Data="M4,4 L12,4 L12,12 L4,12 Z M6,6 L10,6 L10,10 L6,10 Z"
+                              Width="16" Height="16" Foreground="LightBlue"/>
+                    <TextBlock Text="Group" Foreground="LightBlue"/>
+                </StackPanel>
+            </Button>
+
+            <Button Command="{Binding UngroupCommand}"
+                    Width="95" Margin="2" Background="#3d3d3d" ToolTip.Tip="Ungroup (Ctrl+Shift+G)">
+                <StackPanel Orientation="Horizontal" Spacing="4">
+                    <PathIcon Data="M4,4 L6,4 L6,6 L4,6 Z M10,4 L12,4 L12,6 L10,6 Z M4,10 L6,10 L6,12 L4,12 Z M10,10 L12,10 L12,12 L10,12 Z"
+                              Width="16" Height="16" Foreground="Orange"/>
+                    <TextBlock Text="Ungroup" Foreground="Orange"/>
+                </StackPanel>
+            </Button>
+
+            <Separator Width="1" Background="Gray" Margin="10,5"/>
+
+            <!-- File Operations -->
+            <Button Content="📄" Command="{Binding NewProjectCommand}"
+                    Width="36" Margin="2" Background="#3d3d3d" ToolTip.Tip="New project (Ctrl+N)"/>
+            <Button Content="📂" Command="{Binding LoadDesignCommand}"
+                    Width="36" Margin="2" Background="#3d3d3d" ToolTip.Tip="Load design"/>
+            <Button Content="💾" Command="{Binding SaveDesignCommand}"
+                    Width="36" Margin="2" Background="#3d3d3d" ToolTip.Tip="Save design (Ctrl+S)"/>
+            <Button Content="💾+" Command="{Binding SaveDesignAsCommand}"
+                    Width="50" Margin="2" Background="#3d3d3d" ToolTip.Tip="Save design as..."/>
+            <Button Content="🐍 Export" Command="{Binding ExportNazcaCommand}"
+                    MinWidth="75" Margin="2" Padding="8,4" Background="#3d5d3d" ToolTip.Tip="Export to Nazca Python"/>
+
+            <Separator Width="1" Background="Gray" Margin="10,5"/>
+
+            <!-- PDK -->
+            <Button Content="📦 Load PDK" Command="{Binding LoadPdkCommand}"
+                    MinWidth="95" Margin="2" Padding="8,4" Background="#3d3d5d" ToolTip.Tip="Load PDK component library"/>
+            <Button Content="❓" Command="{Binding OpenPdkHelpCommand}"
+                    Width="28" Margin="0,2,2,2" Padding="2,4" Background="#3d3d5d" ToolTip.Tip="PDK JSON format help and AI assistance guide"/>
+
+            <Separator Width="1" Background="Gray" Margin="10,5"/>
+
+            <!-- Simulation -->
+            <Button Content="[L] Simulate" Command="{Binding RunSimulationCommand}"
+                    Width="90" Margin="2" Background="#3d5d3d" ToolTip.Tip="Run light simulation (L). Press P to toggle power overlay. Overlay auto-updates on circuit changes."/>
+        </StackPanel>
+    </ScrollViewer>
+
+</UserControl>
